requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
urllib3>=2.0.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
bcrypt>=4.1.0
psycopg[binary]>=3.1.0
